# api/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: MetalDocs API
  version: 1.0.0
  description: >
    API inicial do MetalDocs (v1). Escopo atual sem integracao com IA.
servers:
  - url: http://192.168.0.3:8080/api/v1
paths:
...
  /iam/users:
    get:
      summary: Lista usuarios internos com estado de autenticacao
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListManagedUsersResponse'
    post:
      summary: Cria usuario interno com credencial local inicial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateManagedUserRequest'
      responses:
        '201':
          description: Usuario criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateManagedUserResponse'
...
  /document-types:
    get:
      summary: Lista os tipos documentais suportados
      parameters:
      responses:
        '200':
          description: Lista de tipos documentais
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListDocumentTypesResponse'
        '401':
          description: Nao autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '403':
          description: Sem permissao
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /access-policies:
    get:
      summary: Lista policies de acesso por recurso
      parameters:
        - name: resourceScope
          in: query
          required: true
          schema:
            type: string
            enum: [document, document_type, area]
        - name: resourceId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAccessPoliciesResponse'
        '400':
          description: Requisicao invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '401':
          description: Nao autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '403':
          description: Sem permissao
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
    put:
      summary: Substitui policies de acesso de um recurso
      parameters:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceAccessPoliciesRequest'
      responses:
        '200':
          description: Policies substituidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReplaceAccessPoliciesResponse'
        '400':
          description: Requisicao invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '401':
          description: Nao autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '403':
          description: Sem permissao
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /documents:
    post:
      summary: Cria documento e primeira versao
      parameters:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '201':
          description: Documento criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentCreatedResponse'
        '400':
          description: Requisicao invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '401':
          description: Nao autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '403':
          description: Sem permissao
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
    get:
      summary: Lista documentos
      parameters:
      responses:
        '200':
          description: Lista de documentos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListDocumentsResponse'
        '401':
          description: Nao autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '403':
          description: Sem permissao
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
...
  /search/documents:
    get:
      summary: Pesquisa documentos por filtros
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: ownerId
          in: query
          required: false
          schema:
            type: string
        - name: documentType
          in: query
          required: false
          schema:
            type: string
            enum: [policy, procedure, work_instruction, contract, supplier_document, technical_drawing, certificate, report, form, manual]
        - name: businessUnit
          in: query
          required: false
          schema:
            type: string
        - name: department
          in: query
          required: false
          schema:
            type: string
        - name: classification
          in: query
          required: false
          schema:
            type: string
            enum: [PUBLIC, INTERNAL, CONFIDENTIAL, RESTRICTED]
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [DRAFT, IN_REVIEW, APPROVED, PUBLISHED, ARCHIVED]
        - name: tag
          in: query
          required: false
          schema:
            type: string
        - name: expiryBefore
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: expiryAfter
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Lista filtrada de documentos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchDocumentsResponse'
        '400':
          description: Requisicao invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '401':
          description: Nao autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '403':
          description: Sem permissao
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
...
components:
...
  schemas:
...
    ManagedUserItem:
      type: object
      required: [userId, username, displayName, isActive, mustChangePassword, failedLoginAttempts, roles, createdAt, updatedAt]
      properties:
        userId:
          type: string
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        isActive:
          type: boolean
        mustChangePassword:
          type: boolean
        failedLoginAttempts:
          type: integer
        lockedUntil:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
        roles:
          type: array
          items:
            type: string
            enum: [admin, editor, reviewer, viewer]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ListManagedUsersResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ManagedUserItem'
    CreateManagedUserRequest:
      type: object
      required: [username, displayName, password, roles]
      properties:
        userId:
          type: string
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        password:
          type: string
        roles:
          type: array
          items:
            type: string
            enum: [admin, editor, reviewer, viewer]
    CreateManagedUserResponse:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
...
    CreateDocumentRequest:
      type: object
      required:
        - title
        - documentType
        - ownerId
        - businessUnit
        - department
      properties:
        title:
          type: string
          minLength: 1
        documentType:
          type: string
          enum: [policy, procedure, work_instruction, contract, supplier_document, technical_drawing, certificate, report, form, manual]
        ownerId:
          type: string
          minLength: 1
        businessUnit:
          type: string
          minLength: 1
        department:
          type: string
          minLength: 1
        classification:
          type: string
          enum: [PUBLIC, INTERNAL, CONFIDENTIAL, RESTRICTED]
          default: INTERNAL
        tags:
          type: array
          items:
            type: string
        effectiveAt:
          type: string
          format: date-time
        expiryAt:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
        initialContent:
          type: string
    DocumentCreatedResponse:
      type: object
      required:
        - documentId
        - version
        - status
        - documentType
      properties:
        documentId:
          type: string
        version:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [DRAFT, IN_REVIEW, APPROVED, PUBLISHED, ARCHIVED]
        documentType:
          type: string
    DocumentListItem:
      type: object
      required: [documentId, title, documentType, ownerId, businessUnit, department, classification, status, tags]
      properties:
        documentId:
          type: string
        title:
          type: string
        documentType:
          type: string
        ownerId:
          type: string
        businessUnit:
          type: string
        department:
          type: string
        classification:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        effectiveAt:
          type: string
          format: date-time
        expiryAt:
          type: string
          format: date-time
    ListDocumentsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DocumentListItem'
    DocumentTypeItem:
      type: object
      required: [code, name, description, reviewIntervalDays]
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        reviewIntervalDays:
          type: integer
          minimum: 1
    ListDocumentTypesResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DocumentTypeItem'
    AccessPolicyItem:
      type: object
      required: [subjectType, subjectId, resourceScope, resourceId, capability, effect]
      properties:
        subjectType:
          type: string
          enum: [user, role, group]
        subjectId:
          type: string
        resourceScope:
          type: string
          enum: [document, document_type, area]
        resourceId:
          type: string
        capability:
          type: string
          enum: [document.create, document.view, document.edit, document.upload_attachment, document.change_workflow, document.manage_permissions]
        effect:
          type: string
          enum: [allow, deny]
    AccessPolicyWriteItem:
      type: object
      required: [subjectType, subjectId, capability, effect]
      properties:
        subjectType:
          type: string
          enum: [user, role, group]
        subjectId:
          type: string
        capability:
          type: string
          enum: [document.create, document.view, document.edit, document.upload_attachment, document.change_workflow, document.manage_permissions]
        effect:
          type: string
          enum: [allow, deny]
    ListAccessPoliciesResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AccessPolicyItem'
    ReplaceAccessPoliciesRequest:
      type: object
      required: [resourceScope, resourceId, policies]
      properties:
        resourceScope:
          type: string
          enum: [document, document_type, area]
        resourceId:
          type: string
        policies:
          type: array
          items:
            $ref: '#/components/schemas/AccessPolicyWriteItem'
    ReplaceAccessPoliciesResponse:
      type: object
      required: [resourceScope, resourceId, replacedCount]
      properties:
        resourceScope:
          type: string
        resourceId:
          type: string
        replacedCount:
          type: integer
          minimum: 0
...
    SearchDocumentItem:
      type: object
      required: [documentId, title, documentType, ownerId, businessUnit, department, classification, status, tags, createdAt]
      properties:
        documentId:
          type: string
        title:
          type: string
        documentType:
          type: string
        ownerId:
          type: string
        businessUnit:
          type: string
        department:
          type: string
        classification:
          type: string
          enum: [PUBLIC, INTERNAL, CONFIDENTIAL, RESTRICTED]
        status:
          type: string
          enum: [DRAFT, IN_REVIEW, APPROVED, PUBLISHED, ARCHIVED]
        tags:
          type: array
          items:
            type: string
        effectiveAt:
          type: string
          format: date-time
        expiryAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    SearchDocumentsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SearchDocumentItem'
...
    ApiErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message, details, trace_id]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        trace_id:
          type: string